11/21 A

+++ script/script.js
@@ -7,7 +7,7 @@ const balance2 = document.getElementById("savings-balance");
 const savingsBar = document.getElementById("savings-bar");
 const savingsFill = document.getElementById("savings-fill");
 
-const interestRate = 0.15/365;
+const interestRate = 0.15 / 365;
 
 var bankBalance = 100;
 var savingsBalance = 100;
@@ -17,23 +17,23 @@ let goalSavings = 1000;
 
 let purchaseCount = 0;
 
-if(localStorage.getItem('purchaseCount')>0){
+if (localStorage.getItem('purchaseCount') > 0) {
     purchaseCount = localStorage.getItem('purchaseCount');
 }
-else{
+else {
     localStorage.setItem('purchaseCount', 0);
 }
 
 let intSum = 0;
-if(localStorage.getItem('intSum')>0){
+if (localStorage.getItem('intSum') > 0) {
     intSum = localStorage.getItem('purchaseCount');
 }
-else{
+else {
     localStorage.setItem('intSum', 0);
 }
 
 let rightSplit = 0;
-if(localStorage.getItem('rightSplit')>0){
+if (localStorage.getItem('rightSplit') > 0) {
     rightSplit = localStorage.getItem("rightSplit");
 }
 else {
@@ -41,7 +41,7 @@ else {
 }
 
 let depCount = 0;
-if(localStorage.getItem('depCount')>0){
+if (localStorage.getItem('depCount') > 0) {
     rightSplit = localStorage.getItem("depCount");
 }
 else {
@@ -51,31 +51,31 @@ else {
 //POPUP variables
 
 // initialization (local storage)
-if(localStorage.getItem('balance')!=100) {
+if (localStorage.getItem('balance') != 100) {
     bankBalance = localStorage.getItem('balance');
 }
 else {
-    localStorage.setItem('balance', 100.0); 
+    localStorage.setItem('balance', 100.0);
 }
-if(localStorage.getItem('savingsBalance')!=100) {
+if (localStorage.getItem('savingsBalance') != 100) {
     savingsBalance = localStorage.getItem('savingsBalance');
 }
 else {
-    localStorage.setItem('savingsBalance', 100.0); 
+    localStorage.setItem('savingsBalance', 100.0);
 }
 
 //these are just to make my life easier
-if(localStorage.getItem('balance')==0) {
+if (localStorage.getItem('balance') == 0) {
     localStorage.setItem('balance', 50);
 }
-if(isNaN(localStorage.getItem('balance'))){
+if (isNaN(localStorage.getItem('balance'))) {
     localStorage.setItem('balance', 50);
 }
-if(isNaN(localStorage.getItem('savingsBalance'))){
+if (isNaN(localStorage.getItem('savingsBalance'))) {
     localStorage.setItem('savingsBalance', 50);
 }
 
-function updateAll(){
+function updateAll() {
     updateBalance();
     updateSavings()
     updateSpendingsBar();
@@ -91,10 +91,9 @@ updateAll();
 // console.log(savingsBalance)
 
 
-
 // SPENDINGS FUNCTIONS
 function increaseBalance(amount) {
-    localStorage.getItem('balance')+= amount;
+    localStorage.getItem('balance') += amount;
     balance.textContent = "$" + localStorage.getItem('balance');
     bankBalance = localStorage.getItem('balance')
     updateSpendingsBar();
@@ -115,7 +114,7 @@ function updateBalance() {
 
 // SAVINGS FUNCTIONS
 function increaseSavings(amount) {
-    localStorage.getItem('savingsBalance')+= amount;
+    localStorage.getItem('savingsBalance') += amount;
     balance2.textContent = "$" + localStorage.getItem('savingsBalance');
     savingsBalance = localStorage.getItem('savingsBalance')
     updateSavingsBar();
@@ -123,7 +122,7 @@ function increaseSavings(amount) {
 
 function updateSavingsBar() {
     const progress2 = (savingsBalance / goalSavings) * 100;
-    savingsFill.style.width = progress2+ "%";
+    savingsFill.style.width = progress2 + "%";
 
 }
 
@@ -135,11 +134,11 @@ function updateSavings() {
 
 
 // POPUP
-if (document.body.className.match("indexP")){
+if (document.body.className.match("indexP")) {
     const popUpShown = sessionStorage.getItem("popUpShown");
     const closePopup = document.getElementById("close-popup");
 
-    document.addEventListener("DOMContentLoaded", function() {
+    document.addEventListener("DOMContentLoaded", function () {
         const popup = document.getElementById("popup");
         const closePopup = document.getElementById("close-popup");
 
@@ -147,39 +146,39 @@ if (document.body.className.match("indexP")){
             popup.style.display = "block";
         }
 
-        closePopup.addEventListener("click", function() {
+        closePopup.addEventListener("click", function () {
             popup.style.display = "none";
             // localStorage.setItem("popupClosed", "true");
         });
     });
 
-    closePopup.addEventListener("click", function() {
+    closePopup.addEventListener("click", function () {
         popup.style.display = "none";
         // sessionStorage.setItem("popUpShown", "true");
     });
 
     const daysInput = document.getElementById("daysInput");
 
-    calculateEarnings.addEventListener("click", function() {
+    calculateEarnings.addEventListener("click", function () {
         const days = parseInt(daysInput.value, 10);
         const earnings = days * 20;
-        
+
         var rawInterest = parseFloat(localStorage.getItem(('savingsBalance'))) * interestRate * days;
         const interest = rawInterest.toFixed(2);
 
         intSum = parseFloat(intSum) + interest;
         localStorage.setItem('intSum', intSum);
 
-        if(days>0) {
+        if (days > 0 && parseFloat(localStorage.getItem('savingsBalance'))>0) {
             alert(`Your Savings has increased by $${interest} from interest.`)
         }
 
         localStorage.setItem('savingsBalance', parseFloat(localStorage.getItem('savingsBalance')) + (interest));
 
-        if (days==1){
+        if (days == 1) {
             alert(`You've earned $20 in 1 day from working!`);
         }
-        else if(days>1){
+        else if (days > 1) {
             alert(`You've earned $${earnings} in ${days} days from working!`);
         }
         else {
@@ -187,23 +186,25 @@ if (document.body.className.match("indexP")){
             popup.style.display = "none";
         }
 
-        const savingsPercentage = parseFloat(prompt("What percentage of your earnings would you like to allocate to savings?"));
-        const savings = (savingsPercentage / 100) * earnings;
+        if(days > 0){
+            const savingsPercentage = parseFloat(prompt("What percentage of your earnings would you like to allocate to savings?"));
+            const savings = (savingsPercentage / 100) * earnings;
 
-        if (!isNaN(savingsPercentage)) {
-            if(savingsPercentage==20){
-                alert(`You've allocated $${savings} (${savingsPercentage}% of your earnings) to savings. Nice job! 20% is the recommended amount to allocate to savings.`);
+            if (!isNaN(savingsPercentage)) {
+                if (savingsPercentage == 20) {
+                    alert(`You've allocated $${savings} (${savingsPercentage}% of your earnings) to savings. Nice job! 20% is the recommended amount to allocate to savings.`);
 
-                rightSplit = parseFloat(rightSplit) + 1;
-                localStorage.setItem('rightSplit', rightSplit);
-            }
-            else{
-                alert(`You've allocated $${savings} (${savingsPercentage}% of your earnings) to savings. Keep in mind that it is recommended to allocate 20% of your earnings to savings!`);
-            }
+                    rightSplit = parseFloat(rightSplit) + 1;
+                    localStorage.setItem('rightSplit', rightSplit);
+                }
+                else {
+                    alert(`You've allocated $${savings} (${savingsPercentage}% of your earnings) to savings. Keep in mind that it is recommended to allocate 20% of your earnings to savings!`);
+                }
+            } 
+            localStorage.setItem('savingsBalance', parseFloat(localStorage.getItem('savingsBalance')) + (savings));
+            localStorage.setItem('balance', parseFloat(localStorage.getItem('balance')) + (earnings - savings));
         }
 
-        localStorage.setItem('savingsBalance', parseFloat(localStorage.getItem('savingsBalance')) + (savings));
-        localStorage.setItem('balance', parseFloat(localStorage.getItem('balance')) + (earnings-savings));
 
         popup.style.display = "none";
 
@@ -214,10 +215,10 @@ if (document.body.className.match("indexP")){
 
 // BANK
 
-function depositMoney(){
+function depositMoney() {
     const depositAmt = parseFloat(prompt("How much money would you like to deposit to Savings?"));
 
-    if(depositAmt <= localStorage.getItem('balance')){
+    if (depositAmt <= localStorage.getItem('balance')) {
         localStorage.setItem('savingsBalance', parseFloat(localStorage.getItem('savingsBalance')) + (depositAmt));
         localStorage.setItem('balance', parseFloat(localStorage.getItem('balance')) - (depositAmt));
 
@@ -227,23 +228,23 @@ function depositMoney(){
         updateAll();
         alert(`You've deposited $${depositAmt}. Your Savings balance is now $${localStorage.getItem('savingsBalance')}`);
     }
-    
+
     else {
         alert("That is not a valid amount. Please try again.")
     }
 }
 
-function withdrawMoney(){
+function withdrawMoney() {
     const withdrawAmt = parseFloat(prompt("How much money would you like to withdraw from Savings?"));
 
-    if(withdrawAmt <= localStorage.getItem('savingsBalance')){
+    if (withdrawAmt <= localStorage.getItem('savingsBalance')) {
         localStorage.setItem('savingsBalance', parseFloat(localStorage.getItem('savingsBalance')) - (withdrawAmt));
         localStorage.setItem('balance', parseFloat(localStorage.getItem('balance')) + (withdrawAmt));
 
         updateAll();
         alert(`You've withdrawn $${withdrawAmt}. Your Spendings balance is now $${localStorage.getItem('balance')}`);
     }
-    
+
     else {
         alert("That is not a valid amount. Please try again.")
     }
@@ -261,7 +262,7 @@ function updateFills() {
 // SHOP
 
 // SHOP variables
-if (document.body.className.match("shopP")){
+if (document.body.className.match("shopP")) {
 
     const wnFill = document.getElementById("wn-fill");
 
@@ -269,25 +270,25 @@ if (document.body.className.match("shopP")){
     var needCount = 0;
 
 
-    if(parseFloat(localStorage.getItem('wantCount'))!=0){
+    if (parseFloat(localStorage.getItem('wantCount')) != 0) {
         wantCount = parseFloat(localStorage.getItem('wantCount'));
     }
-    else{
+    else {
         localStorage.setItem('wantCount', 0)
     }
-    if(parseFloat(localStorage.getItem('needCount'))!=0){
+    if (parseFloat(localStorage.getItem('needCount')) != 0) {
         needCount = parseFloat(localStorage.getItem('needCount'));
     }
-    else{
+    else {
         localStorage.setItem('needCount', 0)
     }
 
 
-    if(typeof wantCount === "string"){
+    if (typeof wantCount === "string") {
         wantCount = 0;
         localStorage.setItem('wantCount', 0)
     }
-    if(typeof needCount === "string"){
+    if (typeof needCount === "string") {
         needCount = 0;
         localStorage.setItem('needCount', 0)
     }
@@ -296,7 +297,7 @@ if (document.body.className.match("shopP")){
         // wnFill.style.width = "10%"
         var a = parseFloat(needCount)
         var b = parseFloat(wantCount)
-        var c = (a/(a+b)) * 100
+        var c = (a / (a + b)) * 100
         wnFill.style.width = c + "%";
 
         console.log(needCount, wantCount)
@@ -308,7 +309,7 @@ if (document.body.className.match("shopP")){
 // PURCHASING (using spendings)
 function purchaseItemW(price) {
     if (localStorage.getItem('balance') >= price) {
-        localStorage.setItem('balance', localStorage.getItem('balance')- price);
+        localStorage.setItem('balance', localStorage.getItem('balance') - price);
         bankBalance = localStorage.getItem('balance')
 
         purchaseCount = parseFloat(purchaseCount) + 1;
@@ -325,10 +326,10 @@ function purchaseItemW(price) {
 
         alert("Item purchased successfully!");
 
-        if((parseFloat(wantCount)/3) > (parseFloat(needCount)/5)){
+        if ((parseFloat(wantCount) / 3) > (parseFloat(needCount) / 5)) {
             alert("Woah! You're spending a little too much money on Wants. Keep in mind that you should be spending 50% of your income on Needs and 30% of your income on Wants.")
         }
-    } 
+    }
 
     else {
         alert("Insufficient funds to purchase this item.");
@@ -337,7 +338,7 @@ function purchaseItemW(price) {
 
 function purchaseItemN(price) {
     if (localStorage.getItem('balance') >= price) {
-        localStorage.setItem('balance', localStorage.getItem('balance')- price);
+        localStorage.setItem('balance', localStorage.getItem('balance') - price);
         bankBalance = localStorage.getItem('balance')
 
         purchaseCount = parseFloat(purchaseCount) + 1;
@@ -353,7 +354,7 @@ function purchaseItemN(price) {
         updateSpendingsBar();
 
         alert("Item purchased successfully!");
-    } 
+    }
 
     else {
         alert("Insufficient funds to purchase this item.");
@@ -361,7 +362,7 @@ function purchaseItemN(price) {
 }
 
 // GOALS
-if (document.body.className.match("goalP")){
+if (document.body.className.match("goalP")) {
     const goal1Fill = document.getElementById("goal1-fill");
     const goal2Fill = document.getElementById("goal2-fill");
     const goal3Fill = document.getElementById("goal3-fill");
@@ -369,9 +370,43 @@ if (document.body.className.match("goalP")){
     const goal5Fill = document.getElementById("goal5-fill");
 
 
-    goal1Fill.style.width = parseFloat(localStorage.getItem('savingsBalance'))/10 + "%";
-    goal2Fill.style.width = parseFloat(localStorage.getItem('purchaseCount'))/3 * 10 + "%";
-    goal3Fill.style.width = parseFloat(localStorage.getItem('intSum'))/10 + "%";
+    goal1Fill.style.width = parseFloat(localStorage.getItem('savingsBalance')) / 10 + "%";
+    goal2Fill.style.width = parseFloat(localStorage.getItem('purchaseCount')) / 3 * 10 + "%";
+    goal3Fill.style.width = parseFloat(localStorage.getItem('intSum')) / 10 + "%";
     goal4Fill.style.width = parseFloat(localStorage.getItem('rightSplit')) * 10 + "%";
     goal5Fill.style.width = parseFloat(localStorage.getItem('depCount')) * 20 + "%";
 }
+
+var toy1img = document.getElementById("toy1");
+
+function addItem(x) {
+    if (!localStorage.getItem('imageAdded')) {
+        // Create new img element
+        var img = document.createElement('img');
+        img.src = x;
+        img.alt = 'Dynamic Image';
+    
+        // Append the image to the target div
+        document.getElementById(targetDiv).appendChild(img);
+    
+        // Set the flag in localStorage
+        localStorage.setItem('imageAdded', 'true');
+
+      } else {
+        console.log('Image has already been added.');
+      }
+ 
+}
+
+function checkAndAddImageOnLoad() {
+    // Set the source of your image
+    var imageSource = 'path_to_your_image.jpg';
+    // Set the ID of your target div
+    var targetDivId = 'targetDiv';
+  
+    // Add image to div if not added before
+    addImageToDiv(imageSource, targetDivId);
+  }
+  
+  // Call the function when the window loads
+  window.onload = checkAndAddImageOnLoad;
